Implement Git hook using yelp/pre-commit
Using the pre-commit framework added various checks to be used
before commits will be accepted.

This resulted in various linting and static issues being resolved,
and the refactoring of the git configuration and setup.  The Git
configuration is now managed solely by the repo itself rather than
coping configuration from the host.