Optimise VM provisioning by skipping/removing steps
The vagrant provisioning was touching too many existing file so
a number of checks have been included to skip when possible (e.g.
if files such as ~/.bash_profile exists).  In addition the GPG key
generation and management has been removed as commit signing
is not currently required - this maybe added back in at a later
date.