Refactor into seperate dev and staging cloud environments
The purpose of this commit is to seperate out the cloud environments
so there is an unrestricted env for developers to test their changes
while ensuring staging is protected.

The only way to update staging is via the CI/CD pipeline in CircleCI

Change in the folder structure in prep for seperating generic
and project specific configuration via Ansible and moving the
GCP project bootstrap scripts into the same folder as the
terraform scripts.

Some additional changes were included here including adding the
CircleCI badge and updating some of the documentation based on
recent changes.