Use golangci to perform extensive linting on golang code
The purpose of this commit was to add a pre-commit hook that
uses golangci to perform a range of linting on the golang codebase.
This will ensure better code is committed to remote.

By introducing as hook this enables the simplification of the
container.sh to focus on container building, scanning, labelling
etc and provides a better separation of concerns.

This will also hugely simplify the CI/CD pipeline as the rules can
use reused across local and automated processes.