Improve k8s management and prepare for addition of reverse proxy
Refactored the account and secret that is required to enable images
to be pulled from GCR to GKE.  This was previously via shell scripts
but this was fragile and would require replicating in the CI/CD
pipeline.  This has now been refactored so that this is accomplished
by Terraform.

In addition Octant has been installed to enable an enhanced UI
experience when working with k8s.  This has been installed using
Ansible as a background service using systemd.

Eventually a reverse proxy will be used so that a single port is needed
to forward HTTPS traffic into the VM.  Caddy 2 has been installed but
is not yet configured.

Further changes have been made to:
1. Fix some defects in the Vagrant provisioning
2. Some minor changes/improvements to the docs
3.

[Ticket: X]