Resolve Terraform authentication documented in Issue #1
The main purpose of this commit is to resolve issue #1 that was encountered
as a result of the Kube context not being set in the CircleCI container.

This is the first commit not being pushed directly to master. Branch protection
has been enabled and a new branch 'develop' created.  In addition Codebeat has
been added as an additional PR check.

A couple of other minor refactoring changes have also been made to the docs.

Issue #1