Improved Vagrant provisioning and basic proxy config
The purpose of commit was to improve the provisioning of the
development VM to install services and resolve some defects.

Octant was configured to run as a service under systemd control
and made available on port 7777 of the VM - however see below.

In addition Caddy was installed and configured as a reverse proxy
for services running inside the VM. However some issues were
ecountered that need to be resolved in future commits.

For more information on these issues please review #5 and #6