Add post-provisioning infrastructure tests with Chef Inspec
This commit introduces an example infrastructure test that is
implemented using Chef Inspec.  This enables tests to be performed
to validate that cloud infrastructure meets acceptance tests.  For
now there is a single compliance check to demonstrate an extensive
set of rules.

This is achieved by configuring the Inspec default files, update
the CircleCI to include the job.

In addition to the above the steps in the job with ugly names were
updated to be user friendly.