Expand the pipline to include infra tests and deployment
Introduced InSpec into the project to enable post provisioning
testing of the infrastructure to to validate while deploying.

Added a step using Skaffold to deploy the application pod into
the staging cluster.  This approach replicates the local env
approach for consistentency.

NOTE: There were some issues with file renaming which was
problematic and resulted in having to reset commits