Change k8s definition from pod to deployment
The main purpose of this commit was to change the example API skaffold
deployment from being a pod and therefore not durable to use a k8s
deployment object.

During the resolution of this it was identified that there there were
issues in loading the configuration of the service which impacted the
quality of the logging that was available, therefore changes to the go
code were applied.

In addition the ingress controller was changed so that the traffic is
redirected into the customer service using port 80 before being forwarded
to 1102.

Also the tagging stratgy used by Skaffold was changed as that enabled the
speeding up of the debugging, this resolved another ticket (#17).

There were some issues encountered with VM cpu availability, so the node
machine size has been parameterised.

Tickets:
https://github.com/delineateio/delineateio.engine/issues/17
https://github.com/delineateio/delineateio.engine/issues/14