Implement isolation in CircleCI pipeline for the .dev & .pub env
The purpose of this change is so that changes to the .pub environment
only happen once a PR is merged and avoid duplicate processing.
Previously CI/CD was running against the .pub environment when the a
change was pushed to any remote branch and then again when the PR was
merged.  That will now not happen.

All pushes on other branches will be applied against the .dev environment
for testing purposes.

Additionally the number of env variables managed within CircleCI has
been reduced for better managability.  Secrets have been relocated to
GCP Secret Manager and dynamically retrieved improving security and
ensuring that secrets being rolled are automatically picked up.

For values that are not secrets these have been moved into the repo so
that they can be managed seperately by environment (e.g. machine size),
are versioned and can be changed by contributors without access to the
CircleCI user interface.

Add printing of env variables to help any debugging!

Add multi line entries