Add check to only deploy software components that have changed (#23)
The main purpose of this commit is to only deploy component changes so that environments only updated if actually necessary. This also has the benefit of speeding up the CircleCI pipeline for the previously slowest step.

This was implemented by creating a new bucket to hold hashes of the component contents so that changes can be checked for for the last available deployed version.

The CircleCI config was updated so that the current hash is read and compared against the new hash and the CircleCI job halted if it is equal. The steps in the job were rearranged so that this operation could be performed before costly installs.

Additional changes were made for scripts...

Script header comments were added to all scripts
Container.sh was renamed to secure.sh which is more descriptive