Apply Github best practices for larger projects (#24)
A CODEOWNERS file was added to control who can make changes to what code.  This is in preparation for additional contributors.

A new SECURITY.md file was added so that the security policy for the project is clear to the community through and is shown on the menu in the GitHub project UI.

Renamed the readme.md to README.md as looking at other projects this appears to be the most common casing.

There was a missing instruction in the README.md file around the local development environment setup related to the GCP service account key location - the opportunity was taken to add this.