Add security policy to services using Cloud Armor
The ultimate purpose of this change is to add network security to
the API endpoints.

This has been implemented by creating a security policy using Terraform
that is not associated to a specific target.  Chanages were made to the
k8s service definition and adding of a k8s BackendConfig CRD to associate
the policy to the services (and therefore pods).

The initial policy implemented policy restricts traffic to only be recieved
from the 14 Cloudflare IP addresses.  The list of Cloudflare IP are here:

https://www.cloudflare.com/ips-v4

In a future PR this list will be changed to be retrieved dynamically.

As part of this commit the opportunity to perform two minor refactorings
was taken:

* `infra` directory was renamed to `cluster`
* gcloud image version was bumped from 300.0.0 to 302.0.0
* Terraform image version was bumped from 0.12.28 to 0.12.29

Tickets:
https://github.com/delineateio/delineateio.engine/issues/25