Abstract the customers service from the platform repo
The purpose of the change is to abstract the customer service out from
the platform code.  This will enable services to be managed separately
from the main platform.

This has been implemented by using submodules to creates pointers to
other repos that implement services.  CircleCI config was changed so that
submodules are initalised and pulled before the deployments are applied to
the k8s cluster.

For now they will share the same pipeline although this will potentially
change in the future if/when the platform scales up.