Reduce cloud costs and changed naming from 'engine' to 'platform'
Introduced a scheduled destroy workflow in CircleCI so that non-master
branches infrastructure is torn down frequently to a) ensure that environments
can be recreated b) cloud costs are reduced.

Bootstrap.sh script was run to create a new environment from scratch.  This
identified two main changes.

Firstly the creation of the bucket was moved into Terraform so that when the
scheduled destroy runs the footprint of deployments are also destroyed, without
them being destroyed then deployments are not triggered.  While testing the
script it highlighted that some of the steps for removing default network were
not being automatically being accepted.

As part of the destroy and recreate workflow testing was to performed to validate
that imaging pulling from GCR was working.

The display names and descriptions were not be created for the infrastructure
account (via bootstrap.sh) and registry (via Terraform) service accounts.  For
consistency both names and descriptions were added.

While testing a warning was encountered when Snyk was run warning of that v8 of
Node was being used but that was no longer supported.  Changes were made to the
Ansible provisioning of the VM to install NodeJS v10.