Move network & policy creation into their own CI/CD step in CircleCI (#31)
The purpose of this change is to isolate network and policy related provisioning into a dedicated step so that these resources are not destroyed when environments are scheduled to be torn down.  There is minimal cost implications of now removing the network along with the other resources.

This was implemented by removing the network and policy related sources from the cluster and ingress Terraform files respectively.  To do this some extra data resources add given the absence of the resources to reference.

As part of implementing the security policy a fixed list of Cloudflare IPs were being used to configure the Cloud Armor policy.  This has now been refactored so that the IPs are dynamically retrieved.  The benefit is that if there are any changes in the required IPs then changing the policy is fully automated.

There were some other minor refactoring that happened, specifically:

* The configure files across Terraform were combined by env
* Names of the CircleCI jobs were renamed so they are aligned to the Terraform lifecycle