Expand the set of infrastructure tests and improve paramterisation
The main purpose of this commit was to expand the basic testing to cover
the network & storage tests and some additional basic cluster tests.

There were challenges testing locally as the dev environment was not
up and running.  Therefore the scjheduling of the env destroy has been changed
aligned to working patterns.