Expand the set of infrastructure tests and improve parametrisation (#32)
The main purpose of this commit was to expand the basic testing to cover the network & storage tests and some additional basic cluster tests.

There were challenges testing locally as the dev environment was not up and running.  Therefore the scheduling of the env destroy has been changed aligned to working patterns.