Add external DB connectivity from the cluster to GCP Cloud SQL
The main purpose of this change is to move the data storage outside
the cluster so it becomes permanent storage.

From an implementation perspective this mean't:
- Introducing a new CircleCI job specifically for the Terraform DB creation
- Add a new CircleCI job for linking the service configuration to the pod
- Implemented Workload Identity for the cluster and pods (best pratice to grant db access)
- Automated the creation of DB related secrets into k8s using tf
- The .env files were moved to improve the project structure aligned to the next point
- Moved the .tfvars files to a common location as many of the inputs are shared across jobs
- Domain name was moved from Secrets Manager to a tf variable to simplify tf code
- Corresponding changes in the submodule service so the reference commit has been bumped

During the testing of this solution it was identified that the resource limits of the
containers in the pod was too high.  The consequence of this was that the machine size
had been oversized / could be reduced and thus was increasing costs.  The container have
now been limited further enabling the node machine sizes to be reduced.

Implemented a new connect.sh script that helps to reconnect to a cluster if the cluster
has been destroyed as part of the terraform changes.