Fix issue encountered in Database commit (424dbb1) on the Workload SA
The primary purpose of this commit is to fix an issue encountered in a previous commit
whereby the role assignment to the 'workload' service account which is used as the workload
identity of the pods had not been implemented.  When encoutered it was possible to manually
assign the roles in the GCP console to prove the fix.

The roles that are required by the pod are as follows.  Without these the pods failed to start
cleanly as the permissions where denied when starting the SQL Proxy

* Cloud SQL Client
* Logs Writer
* Monitoring Metric Writer
* Monitoring Viewer