Fix issue encountered during environment promotion for Database commit (424dbb1) (#36)
During the promotion of the Database commit an issue was encountered because the required services for .pub had changed from when the project was boot strapped. As an example the Cloud SQL API was not enabled so the DB creation failed.

To resolve this issue the a new job in CircleCI that executes before the network job enables the required services, the bootstrap.sh was edited to remove the service enablement. It should be noted that the Service Usage API is still required to be enabled before tf is executed in the pipeline.