Add BDD based API tests into the pipeline for the example API
**Purpose**
Ensure that there is a robust testing approach in place that enables us ongoing
integration testing of the changes that are being made.

**Approach**
Use the Python Behave framework for BDD and implement step definitions that are
based on the concise DSL implemented in the Karate framework

**Refactoring**
There was a signficant amount of refactoring within this commit including but not
only:

The implementation of IAM assigment to the workload identity was reafactored as this
was destructive and kept losing the required permissions.

Refinement of the VM provisioning scripts to remove unused components, most notably
Microk8s as that just become an additional workflow step and is unneeded with the dev
environment

Refactoring of the CircleCI config file to accomodate the test job but also remove the
repeating steps into the custom commands for Terraform etc.

**Tickets**
None