Improve the local developer experience (#38)
**Purpose**

Refactor the solution to improve the developer experience when working locally on the machine.

**Implementation**

Refactoring the Ansible scripts used by Vagrant to remove unrequired steps so that provisioning speed is improved. VMWare Fusion as unstable as a provider as the SSH speed slowed down dramatically while port collisions were becoming increasingly common.

Rename the `core` directory to `install` to be align to the two step process. In addition defects in the Docker install and management were resolved while refactoring out the use of snap for installs.

Changes were made so the API tests can be run on the host, in VM and in the CircleCI pipeline.