Move the tf config from service to platform
**Purpose**

Part of the overall strategy to slim down the service to the minimum
so an individual service has less artefacts

**Approach**

Move the tf files to be another component of the cloud config and call
the `terraform apply` against the platform rather than service directory.

**Additional**

In addition to this a new destroy.sh script was adding enabling the
environments to be taken down.  Also generated passwords were changed
to not include special chars as this seemed to cause some issues with
the construction of the database string.

There is potentially an issue with the provisioning when the clsuter is
being rebuilt.  If this turns out to be true this will be handled by a
another PR.