Migrated to delineateio/box v1.0.1 Vagrant box
**Purpose**

The purpose of this change is to minimise the custom implementation that resides within the project.

**Approach**

Refactor the project by updating the `vagrantfile` and removed the unnecessary Ansible and configuration files.  The base box being used in Vagrant had Terraform v0.14 and this caused issues in terms of providers.

* Refactored and removed the setup now encapsulated in `box`
* Downgrade the version of Terraform in the VM
* Apply version contraints in the terraform blocks
* Bumped other versions to be consistent across local & CI/CD
* Upgraded the pre-commit versions to improve checks

**Refactoring**

Refactored the location of the .gcloud service account key.

**Tickets**

None